Make use of the new argument parser
