+++ Package.swift
@@ -16,12 +16,12 @@ let package = Package(
     dependencies: [
         // dev .package(url: "https://github.com/danger/swift", from: "3.0.0"),
         // dev .package(path: "Submodules/WeTransfer-iOS-CI/Danger-Swift"),
-        .package(url: "https://github.com/apple/swift-package-manager.git", from: "0.1.0")
+        .package(url: "https://github.com/apple/swift-argument-parser", from: "0.0.1"),
     ],
     targets: [
         // dev .testTarget(name: "PoesTests", dependencies: ["PoesCore"]),
         // dev .target(name: "DangerDependencies", dependencies: ["Danger", "WeTransferPRLinter"], path: "Submodules/WeTransfer-iOS-CI/Danger-Swift", sources: ["DangerFakeSource.swift"]),
         .target(name: "Poes", dependencies: ["PoesCore"]),
-        .target(name: "PoesCore", dependencies: ["SPMUtility"])
+        .target(name: "PoesCore", dependencies: ["ArgumentParser"])
     ]
 )
