+++ Poes.xcodeproj/xcshareddata/xcschemes/Poes.xcscheme
@@ -29,16 +29,6 @@
       shouldUseLaunchSchemeArgsEnv = "YES"
       codeCoverageEnabled = "YES">
       <Testables>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "Poes::PoesTests"
-               BuildableName = "PoesTests.xctest"
-               BlueprintName = "PoesTests"
-               ReferencedContainer = "container:Poes.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
