+++ Sources/Poes/main.swift
@@ -1,7 +1,3 @@
 import PoesCore
 
-do {
-    try Poes.run()
-} catch {
-    print("Pushing failed: \(error)")
-}
+Poes.main()
